requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.3
gunicorn==21.2.0
pydantic==2.8.2
pydantic-settings==2.3.4
# Removed SQLAlchemy, alembic, and psycopg2 - using MongoDB only
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.8
celery==5.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

